Bug 518860 - Generator: rename the class name of OSLC Services to be based on the oslc:usage property
Added zero-or-many usage property on each of the Service,
CreationFactory, QueryCapability and Dialogs properties. the
corresponding methods are now annotated accordingly.

Also, renamed the Service class (as well as each of the capabilities jsp
files) to be based on the usage property, as
follows:

* If no oslc:usage property is defined, the service class name is based
on the name of its SP, followed by an index. (for example,
ServiceProviderService1.java).

* If there are more than 1 usage property defined, the service class
name is based on the first oslc:usage property value. The Service class
name is the last token from the usage URI after tokenizing the
property using the characters "#", "/" and "\"

Change-Id: I604cfcf5267110ef40a4ca8a906dfd28c81ee2ae
Signed-off-by: Jad <jad@kth.se>