Bug 519663 - Allow DomainSpecifications to override the default packagenames and locations of their resulting javaclasses
A GenerationSettings property can be set on the whole Specification
and/or a specific DomainSpecification. This settings
allows the modeller to override the default path & package names defined
for a specific adaptorInterface.

For a Domain Specification, The GenerationSettings property are modified
through a context menu on a Domain Specification element in the
Modeller.
For a Specification, the GenerationSettings property are modified
through a context menu on Specification diagram in the Modeller.

For each Domain specification (or for the whole Specification), its
resources and a "Constants" interface
class is generated at its destination path (default or overrriden).

For each adaptor, a resourceCreationFactory class is generated to
support the construction of resource instances (or their URI/links) that
is specific for that particular adaptor.
The adaptor-specific constructors and ConstructURI/Link methods
originally generated in the resource classes are now marked as
deprecated.

Change-Id: I53d3378d75b19025ad2a45e05231673204ecbabb
Signed-off-by: Jad <jad@kth.se>