Bug 521222 - Generator: generate code per-adaptor
Introduced a context menu on AdaptorInterface elements that allows user
to trigger generation for that specific adaptorInterface.

The user is then prompted for the base path (unlike the current
generation plugin that expects the paths to be relative to the model
file).

Also, the user can create a generator.properties file with the base
path, avoiding the repeated prompt for path.

Wiki pages to be updated once commit is merged.

Change-Id: Ia83f3eddcc207afd1020d3bfd4f1de47e9d7b793
Signed-off-by: Jad <jad@kth.se>
Signed-off-by: Andrew Berezovskyi <andriib@kth.se>