Ambiguous methods created for the Creation Factory in the Mgr class (#37)
The problem occurs when there are 1+ capabilities (such as
CreationFactory) within an adaptor that manage the same (set of)
resources. This results in multiple methods in the AdaptorManager that
have the same signature.

To remedy, for such 1+ capabilities, append the name of the capability
to the method. For example, a creation factory - with title CF1 - on
Requirement will result in method called "createRequirementForCF1"

This change also made it necessary to have a additional method in
AdaptorManager to handle the creation method from a CreationDialog
(previously, both CreationDialog and CreationFactory) shared the same
createResource method in Manager.